requirements.txt
Flask
Werkzeug
SQLAlchemy
Flask-SQLAlchemy
psycopg2-binary
GeoAlchemy2
Flask-Migrate
Alembic
Flask-Login
Flask-WTF
WTForms
Pillow
Flask-Mail
python-dotenv
scikit-learn
tensorflow
numpy
pandas
scipy
matplotlib
seaborn
Shapely
Fiona
pyproj
python-dateutil
requests
jsonschema
cryptography
pytest
pytest-flask
Flask-DebugToolbar
gunicorn
reportlab
python-magic
PyPDF2
python-docx
pytesseract
pdf2image
PyMuPDF
